+++ src/main/java/org/c4marathon/assignment/settlement/domain/Settlement.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.settlement.entity;
+package org.c4marathon.assignment.settlement.domain;
 
 import static jakarta.persistence.FetchType.*;
