+++ src/main/java/org/c4marathon/assignment/settlement/domain/SettlementType.java
@@ -1,0 +1,6 @@
+package org.c4marathon.assignment.settlement.domain;
+
+public enum SettlementType {
+	EQUAL,
+	RANDOM
+}
