+++ src/main/java/org/c4marathon/assignment/settlement/domain/repository/SettlementRepository.java
@@ -1,8 +1,8 @@
-package org.c4marathon.assignment.settlement.entity.repository;
+package org.c4marathon.assignment.settlement.domain.repository;
 
 import java.util.List;
 
-import org.c4marathon.assignment.settlement.entity.Settlement;
+import org.c4marathon.assignment.settlement.domain.Settlement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SettlementRepository extends JpaRepository<Settlement, Long> {
