+++ src/main/java/org/c4marathon/assignment/settlement/dto/SettlementRequest.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.settlement.dto;
 
 import java.util.List;
 
-import org.c4marathon.assignment.settlement.entity.SettlementType;
+import org.c4marathon.assignment.settlement.domain.SettlementType;
 
 public record SettlementRequest(
