--- src/main/java/org/c4marathon/assignment/settlement/entity/SettlementType.java
@@ -1,6 +1,0 @@
-package org.c4marathon.assignment.settlement.entity;
-
-public enum SettlementType {
-	EQUAL,
-	RANDOM
-}
